debugging

+++ client_MPath_/include/timer.h
@@ -1,8 +1,8 @@
 
 class Timer {
 public:
-	int num_timeSlice;
-	int len_timeSlice;
+	double num_timeSlice;
+	double len_timeSlice;
 
 	Timer();
